requirements.txt
Django==1.9
argparse==1.2.1
django-datetime-widget==0.9.3
django-widget-tweaks==1.4.1
psycopg2==2.6.1
pytz==2015.7
wsgiref==0.1.2
